==> R/03_Clean_and_Explore_Data.R <==
# -----------------------------------------------------------------------------#
# Syringodium isoetifolium phyloseq cleanup
# Cleaning up the phyloseq object
# Author: Geoffrey Zahn
# Software versions:  R v 4.2.2
#                     tidyverse v 1.3.2
#                     phyloseq v 1.40.0
#                     ShortRead v 1.54.0
#                     Biostrings v 2.64.0
#                     adegenet v 2 .1 10
#                     readxl v 1.4.1
#                     janitor v 2.1.0
#                     microbiome v 1.20.0
# -----------------------------------------------------------------------------#

#################################################################################
#                               Main workflow                                   #
#        Remove non-bacteria, chloroplast and mitochondrial sequences           #
#                                                                               #
#################################################################################

# SETUP ####
library(tidyverse); packageVersion("tidyverse")
library(phyloseq); packageVersion("phyloseq")
library(ShortRead); packageVersion("ShortRead")
library(Biostrings); packageVersion("Biostrings")
library(adegenet); packageVersion("adegenet")
library(readxl); packageVersion("readxl")
library(janitor); packageVersion("janitor")
library(microbiome); packageVersion("microbiome")

source("./R/googlemap_styling.R")
# seed
set.seed(789)

# DATA ####
ps <- readRDS("./output/ps_not-cleaned_w_tree.RDS") # change to non-phylogeny stuff

# EXPLORE ####
ps_nonbact <- subset_taxa(ps, Kingdom != "Bacteria")

# quick plot to look at kingdom-level taxonomy
# ps %>% transform_sample_counts(function(x){x/sum(x)}) %>%
#   plot_bar(fill="Kingdom")
# ggsave("./output/figs/16S_Kingdom-Level_Taxonomic_Proportions.png",dpi=300) # save figure for later use

# same plot, but non-bacteria, for sanity check
# ps_nonbact %>% 
#   transform_sample_counts(function(x){x/sum(x)}) %>%
#   plot_bar(fill="Kingdom")


# REMOVE NON-BACTERIA ####
ps <- subset_taxa(ps, Kingdom == "Bacteria")
tax <- tax_table(ps)

ps <- subset_taxa(ps,Class != "Chloroplast")
ps <- subset_taxa(ps, taxa_sums(ps) > 0)
ps <- subset_samples(ps, sample_sums(ps) > 0)

# Save DNA sequences apart from rownames (from subsetted ps object)
seqs <- taxa_names(ps)
seqs <- DNAStringSet(seqs)
saveRDS(seqs,"./output/16S_ASV_reference_sequences.RDS")


# double-check metadata

# Bangka has both east and west?
bad_bangka <- which(ps@sam_data$east_west == "West" & ps@sam_data$location == "Bangka")
ps@sam_data$east_west[bad_bangka] <- "East" # fix it


# ADD EXTERNAL DATA ####

## seagrass genotypes (geoff) ####
gt <- read_xlsx("./data/SI_GenoTypes_All.xlsx") %>% 
  clean_names()
# clean up sample names to match ps object
sites <- gt$name %>% str_split("(?<=[A-Za-z])(?=[0-9])")
gt$name <- paste0(
  sites %>% map_chr(1),
  "_",
  sites %>% map_chr(2)
) %>% 
  str_replace("_1$","_01") %>% # add leading zeros
  str_replace("_2$","_02") %>% 
  str_replace("_3$","_03") %>% 
  str_replace("_4$","_04") %>% 
  str_replace("_5$","_05") %>% 
  str_replace("_6$","_06") %>% 
  str_replace("_7$","_07") %>% 
  str_replace("_8$","_08") %>% 
  str_replace("_9$","_09") %>% 
  str_replace("Derwan_","Derawan_") # fix differential spelling
# subset to only those that match ps object
gt <- gt[which(gt$name %in% sample_names(ps)), ]
# make genind object
genind <- df2genind(X = select(gt,starts_with("s")),
          ncode = 3,
          ind.names = gt$name,
          ploidy = 1)
# pop clust analysis
clust <- find.clusters(genind,n.pca = 20,max.n.clust = 12,n.clust = 12)
# build metadata
pop_group_metadata <- 
microbiome::meta(ps) %>% 
  full_join(data.frame(sample = names(clust$grp),
                       pop_group = clust$grp))
# add to ps object
x <- sample_data(pop_group_metadata) # convert to sample_data object
sample_names(x) <- x$sample # make sample names match
ps@sam_data <- x # add to ps

## Precipitation data (Josh Leon) ####
#source("./R/Leon_metadata.R")
# add to ps object sample_data


## Sea level - mean monthly (Kate Hickman) ####
source("./R/kh_metadata.R")
sea_level <- read_csv("./data/kh_metadata.csv")
# add to ps object sample_data
new_metadata <- sea_level %>% 
  dplyr::rename(sample = sample_id) %>% 
  full_join(microbiome::meta(ps)) %>% 
  sample_data()
sample_names(new_metadata) <- new_metadata$sample
ps <- phyloseq(otu_table(ps@otu_table),
         tax_table(ps@tax_table),
         phy_tree(ps@phy_tree),
         new_metadata)

## monthly precipitation ####
source("./R/Leon_metadata.R")
precip <- read_csv("./data/precip_metadata.csv")
new_metadata <- precip %>% 
  full_join(microbiome::meta(ps)) %>% 
  sample_data()
sample_names(new_metadata) <- new_metadata$sample
ps <- phyloseq(otu_table(ps@otu_table),
               tax_table(ps@tax_table),
               phy_tree(ps@phy_tree),
               new_metadata)

## elevation ####
elev <- read_csv("./data/elevation_data.csv")
new_metadata <- elev %>%
  dplyr::rename("sample" = "sample_name") %>% 
  full_join(microbiome::meta(ps)) %>% 
  sample_data()
sample_names(new_metadata) <- new_metadata$sample
ps <- phyloseq(otu_table(ps@otu_table),
               tax_table(ps@tax_table),
               phy_tree(ps@phy_tree),
               new_metadata)



# Save RDS object for cleaned up Phyloseq object
saveRDS(ps, file = "./output/clean_phyloseq_object.RDS")
